refactor(0570): move the Z-series collector to _vars_zseries.py
The collector plus its two helpers pushed compute_vars.py to 837 lines, over
the 800-line god-module gate in test_script_hygiene. Adding the three spread
vars is what crossed it, so the split belongs in this ticket rather than in a
follow-up that inherits a red lint.

Follows the seam the house style already uses for a self-contained group of
variables — _vars_registry, _vars_retrieval, _vars_ablation — rather than
inventing a new one. compute_vars drops to 699 lines and imports
zseries_stats; the module owns its own logger and DERIVED_TABLES_DIR, which is
what the tests monkeypatch.

Pure move: the vars file regenerates byte-identically, checked with make -B
and md5sum rather than inferred from a green suite.

Ticket-ref: tickets/0570-a-vars-file-ships-missing-into-a-results.erg
